Fix the implementation of timer_sleep.
Changed the implementation of timer_sleep to eliminate busy waiting.
The fix was implemented by adding a wakeup_time field in the thread struct,
making an ordered list of sleeping threads in timer.c, blocking the thread
calling timer_sleep and adding it to the sleeping_threads list and checking
on each timer_interrupt if there are any threads that should wake up and if so
unblock them and remove them from the sleeping_threads list.